Big Mac Commit! Updating plug-ins to use Release instead of Debug for command line building. Update to use 64-bit safe datatypes (NSInteger where needed). Getting rid of no longer used variables in the Xcode projects. Making a Localizable.strings file.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@72225 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ macosx/PcsxrPlugin.h
@@ -29,7 +29,7 @@
 - (NSString *)getDisplayVersion;
 - (int)getType;
 - (NSString *)path;
-- (unsigned long)hash;
+- (NSUInteger)hash;
 - (NSString *)description;
 - (BOOL)hasAboutAs:(int)type;
 - (BOOL)hasConfigureAs:(int)type;
