+++ macosx/plugins/DFInput/macsrc/ControllerList.h
@@ -34,8 +34,8 @@
 + (void)setCurrentController:(int)which;
 + (int)currentController;
 + (int)getButtonOfRow:(int)row;
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView;
-- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex;
-- (void)deleteRow:(int)which;
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView;
+- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex;
+- (void)deleteRow:(NSInteger)which;
 
 @end
