+++ macosx/plugins/PeopsXgl/macsrc/drawgl.h
@@ -4,7 +4,7 @@ unsigned long ulInitDisplay(void);
 void CloseDisplay(void);
 void BringContextForward(void);
 void DoBufferSwap(void);
-void SetVSync(long myValue);
+void SetVSync(GLint myValue);
 void BringContextForward(void);
 void ChangeWindowMode(void);
 void AboutDlgProc();
